# src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ErrorResponsePayload": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateUserDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserEntity": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"password",
					"createdAt",
					"updatedAt",
					"deletedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TokenResponse": {
				"properties": {
					"accessToken": {
						"type": "string"
					},
					"refreshToken": {
						"type": "string"
					}
				},
				"required": [
					"accessToken",
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginBodyDto": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RefreshTokenDto": {
				"properties": {
					"refreshToken": {
						"type": "string"
					}
				},
				"required": [
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			},
			"swagger_ui_extra_configuration": {
				"persistAuthorization": true
			}
		}
	},
	"info": {
		"title": "Real Estate Ecommerce",
		"version": "1.0.0",
		"description": "One place for all things",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/{teamId}/user": {
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"deletedAt": {
											"type": "string",
											"format": "date-time",
											"nullable": true
										},
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"name": {
											"type": "string"
										},
										"password": {
											"type": "string"
										},
										"email": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"deletedAt",
										"updatedAt",
										"createdAt",
										"name",
										"password",
										"email",
										"id"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "BadRequest",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponsePayload"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"409": {
						"description": "ConflictException",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponsePayload"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"summary": "회원가입",
				"tags": [
					"/user"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "teamId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserDto"
							}
						}
					}
				}
			}
		},
		"/auth/register": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserEntity"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserDto"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "LoginUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TokenResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginBodyDto"
							}
						}
					}
				}
			}
		},
		"/auth/refresh": {
			"post": {
				"operationId": "RefreshToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TokenResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RefreshTokenDto"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:9000/"
		}
	]
}